refactor(progress-reporter): extract composition module (resolves #1861)
